+++ src/pq_ttd.h
@@ -3,12 +3,12 @@
 
 #define PHOTONBLOCK 1024
 
-int pt2_v2_to_ttd(pq_rec_t pq_rec, ttd_t *ttd_rec, ttd_t *overflow_correction, pq_fileinfo_t *file_info);
-int pt3_v2_to_ttd(pq_rec_t pq_rec, ttd_t *ttd_rec, ttd_t *overflow_correction, pq_fileinfo_t *file_info);
-int ht2_v1_to_ttd(pq_rec_t pq_rec, ttd_t *ttd_rec, uint64_t *overflow_correction, pq_fileinfo_t *file_info);
-int ht2_v2_to_ttd(pq_rec_t pq_rec, ttd_t *ttd_rec, uint64_t *overflow_correction, pq_fileinfo_t *file_info);
-int ht3_v1_to_ttd(pq_rec_t pq_rec, ttd_t *ttd_rec, uint64_t *overflow_correction, pq_fileinfo_t *file_info);
-int ht3_v2_to_ttd(pq_rec_t pq_rec, ttd_t *ttd_rec, uint64_t *overflow_correction, pq_fileinfo_t *file_info);
+int16_t pt2_v2_to_ttd(pq_rec_t pq_rec, ttd_t *ttd_rec, ttd_t *overflow_correction, pq_fileinfo_t *file_info);
+int16_t pt3_v2_to_ttd(pq_rec_t pq_rec, ttd_t *ttd_rec, ttd_t *overflow_correction, pq_fileinfo_t *file_info);
+int16_t ht2_v1_to_ttd(pq_rec_t pq_rec, ttd_t *ttd_rec, uint64_t *overflow_correction, pq_fileinfo_t *file_info);
+int16_t ht2_v2_to_ttd(pq_rec_t pq_rec, ttd_t *ttd_rec, uint64_t *overflow_correction, pq_fileinfo_t *file_info);
+int16_t ht3_v1_to_ttd(pq_rec_t pq_rec, ttd_t *ttd_rec, uint64_t *overflow_correction, pq_fileinfo_t *file_info);
+int16_t ht3_v2_to_ttd(pq_rec_t pq_rec, ttd_t *ttd_rec, uint64_t *overflow_correction, pq_fileinfo_t *file_info);
 
 int read_header(FILE *fpin);
